pyproject.toml
[project]
name = "instagram-dms"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Dania Rifki", email = "atomicnumberphi@gmail.com" }
]
dependencies = [
    "pywebview>=5.2",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "pyinstaller>=6.10.0",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/instagram_dms"]
